Reset reconnection count after successful connection
Errbot features builtin reconnection logic which will do exponential
backoff to prevent overloading a network in case of failure. It is
important however to reset the state after a successful connection
because otherwise a second failure will start with a backoff time still
set to what it was the previous time it started reconnecting.